% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cell_coloring.R
\name{add_cell_coloring}
\alias{add_cell_coloring}
\title{Add coloring}
\usage{
add_cell_coloring(cell_positions, color_cells = c("auto", "invisible",
  "positions", "grouping", "gene"), task, grouping_assignment = NULL,
  groups = NULL, gene_oi = NULL, expression_source = "expression")
}
\arguments{
\item{color_cells}{How to color the cells}

\item{grouping_assignment}{Tibble containing the assignment of cells to groups of cells}

\item{groups}{Tibble containing information of the cell groups}

\item{gene_oi}{Gene to plot expression}

\item{expression_source}{Source of the gene expression, defaults to `expression`}
}
\description{
Add coloring
}
