% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_milestone_coloring.R
\name{add_milestone_coloring}
\alias{add_milestone_coloring}
\title{Add colouring to a set of milestones.}
\usage{
add_milestone_coloring(
  milestones = NULL,
  color_milestones = c("auto", "given", "cubeHelix", "Set3", "rainbow")
)
}
\arguments{
\item{milestones}{Tibble containing the column `milestone_id` (character).
If `color_milestones` is set to `"given"`, this tibble should also contain a column
`color` (character), containing colour hex codes (e.g. `"#123456"`).}

\item{color_milestones}{Which palette to use for colouring the milestones
* `auto`: Determine colours automatically. If `color` is already specified in
  milestones tibble, this will be used. Otherwise, the colour scheme is determined by
  `milestone_palette_list$auto`.
* `given`: The `milestones` object already contains a column `color`.
* `cubeHelix`: Use the `rje::cubeHelix()` palette.
* `Set3`: Use the `RColorBrewer::brewer.pal(name = "Set3")` palette.
* `rainbow`: Use the `grDevices::rainbow()` palette.}
}
\value{
A tibble containing the input character column `milestone_id` and a character
  column `color` containing colour hex-codes (e.g. `"#123456"`).
}
\description{
Add colouring to a set of milestones.
}
\examples{
milestones <- data.frame(
  milestone_id = c("A", "B", "C")
)
add_milestone_coloring(milestones)

milestones <- data.frame(
  milestone_id = c("A", "B", "C"),
  color = c("#111111", "#222222", "#333333")
)
add_milestone_coloring(milestones)
}
