% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory_dimred.R
\name{check_or_perform_dimred}
\alias{check_or_perform_dimred}
\title{Perform dimred trajectory if needed}
\usage{
check_or_perform_dimred(object, colour_cells = "milestone",
  colour_milestones = "auto")
}
\arguments{
\item{object}{An object generated by \code{wrap_data} or \code{dimred_trajectory}.}

\item{colour_cells}{How to colour the cells. Must be equal to \code{"milestone"}
or a custom character vector of colours, one for each cell.}

\item{colour_milestones}{How to color the milestones.
Must be one of \code{"auto"} (default), \code{"Set3"}, \code{"rainbow"}, \code{"cubeHelix"},
or a custom character vector of colours, one for each milestone.}
}
\description{
Perform dimred trajectory if needed
}
