% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{dimred_dp}
\alias{dimred_dp}
\title{DP dimensionality reduction}
\usage{
dimred_dp(x, ndim = 3, neigen = NULL)
}
\arguments{
\item{x}{Data matrix}

\item{ndim}{Number of dimensions}

\item{neigen}{Number of eigen values}
}
\description{
DP dimensionality reduction
}
