% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{dimred_ica}
\alias{dimred_ica}
\title{ICA dimensionality reduction}
\usage{
dimred_ica(x, ndim = 3)
}
\arguments{
\item{x}{Data matrix}

\item{ndim}{Number of dimensions}
}
\description{
ICA dimensionality reduction
}
