% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{dimred_mds}
\alias{dimred_mds}
\title{MDS dimensionality reduction}
\usage{
dimred_mds(x, ndim = 3)
}
\arguments{
\item{x}{Data matrix}

\item{ndim}{Number of dimensions}
}
\description{
MDS dimensionality reduction
}
