% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{dimred_pca}
\alias{dimred_pca}
\title{PCA dimensionality reduction}
\usage{
dimred_pca(x, ndim = 3)
}
\arguments{
\item{x}{Data matrix}

\item{ndim}{Number of dimensions}
}
\description{
PCA dimensionality reduction
}
