% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory_dimred.R
\name{dimred_trajectory}
\alias{dimred_trajectory}
\title{Perform dimensionality reduction on a trajectory and the respective samples in order to plot it}
\usage{
dimred_trajectory(traj_object, insert_phantom_edges = TRUE)
}
\arguments{
\item{traj_object}{A trajectory as generated by
\code{\link{wrap_ti_task_data}} or
\code{\link{wrap_prediction_model}}}

\item{insert_phantom_edges}{attempt to plot bifurcating edges correctly automatically}
}
\description{
Perform dimensionality reduction on a trajectory and the respective samples in order to plot it
}
