% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{dimred_tsne}
\alias{dimred_tsne}
\title{tSNE dimensionality reduction}
\usage{
dimred_tsne(x, ndim = 3)
}
\arguments{
\item{x}{Data matrix}

\item{ndim}{Number of dimensions}
}
\description{
tSNE dimensionality reduction
}
