% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynplot.R, R/package.R
\docType{package}
\name{dynplot}
\alias{dynplot}
\alias{dynplot-package}
\title{Create a dynplot with a specified layout}
\usage{
dynplot(dataset, layout = layout_dimred(dataset))
}
\arguments{
\item{dataset}{A dynwrap dataset object, typically containing a trajectory}

\item{layout}{A \code{layout_*} function from dynplot, such as \code{\link[=layout_dimred]{layout_dimred()}} or \code{\link[=layout_graph]{layout_graph()}}}
}
\value{
A ggplot2 object, with the processed data in \code{plot$data} and \code{attr(plot$data, "data")}
}
\description{
Create a dynplot with a specified layout

Plot all the trajectories
}
\examples{
toy <- dyntoy::generate_dataset()
dynplot(dataset) +
  geom_cell_point(aes(colour = select_expression("G1"))) +
  scale_expression_fillcolour() +
  new_scale_fillcolour() +
  geom_trajectory_segments(aes(colour = edge_id))

}
