% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_clean.R
\name{empty_plot}
\alias{empty_plot}
\title{Create an empty plot for spacing}
\usage{
empty_plot()
}
\value{
An empty ggplot2.
}
\description{
Create an empty plot for spacing
}
\examples{
empty_plot()
}
\keyword{plot_helpers}
