% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_tree}
\alias{example_tree}
\title{An example tree dataset}
\format{
An object of class \code{dynwrap::with_prior} (inherits from \code{dynwrap::with_expression}, \code{dynwrap::with_cell_waypoints}, \code{dynwrap::with_trajectory}, \code{dynwrap::data_wrapper}, \code{list}) of length 20.
}
\usage{
example_tree
}
\description{
An example tree dataset
}
\keyword{datasets}
