% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_cell_.R
\name{geom_cell_point}
\alias{geom_cell_point}
\alias{geom_cell_hex}
\title{Plotting cells}
\usage{
geom_cell_point(mapping = NULL,
  data = construct_get_cell_info(mapping), ..., show.legend = NA)

geom_cell_hex()
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes().}

\item{data}{A function created by \code{\link[=construct_get_cell_info]{construct_get_cell_info()}}.}

\item{show.legend}{Whether to show a legend for this geom}
}
\description{
Plotting cells
}
