% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_milestone_.R
\name{geom_milestone_label}
\alias{geom_milestone_label}
\title{Plotting milestones}
\usage{
geom_milestone_label(mapping = NULL, data = get_milestone_info, ...,
  show.legend = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes().}

\item{data}{A function created by \code{\link[=get_cell_info_constructor]{get_cell_info_constructor()}}.}

\item{show.legend}{Whether to show a legend for this geom}
}
\description{
Plotting milestones
}
