% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_trajectory_divergence.R
\name{geom_trajectory_divergence}
\alias{geom_trajectory_divergence}
\title{Plotting the divergence regions of a trajectory}
\usage{
geom_trajectory_divergence(mapping = NULL, ...,
  data = get_divergence_info, show.legend = NA)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{data}{A function to get the information on divergences, typically \code{\link[=get_divergence_info]{get_divergence_info()}}.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}
}
\description{
Plotting the divergence regions of a trajectory
}
