% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_trajectory_segments.R
\name{geom_trajectory_segments}
\alias{geom_trajectory_segments}
\title{Plotting individual segments of the trajectory}
\usage{
geom_trajectory_segments(mapping = NULL,
  position_arrow = position_trajectory_arrows_middle(),
  arrow = ggplot2::arrow(length = ggplot2::unit(0.5, "cm"), type =
  "closed"), arrow_size = 1, shadow = if ("colour" \%in\%
  names(mapping)) {     dynplot::shadow() } else {     FALSE }, ...,
  data = construct_get_segment_info(position_arrow), show.legend = NA)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{position_arrow}{Where to place the arrows within the segments. Typically these are functions created by \code{\link[=position_trajectory_arrows_middle]{position_trajectory_arrows_middle()}} or \code{\link[=position_trajectory_arrows_boundaries]{position_trajectory_arrows_boundaries()}}.}

\item{arrow}{specification for arrow heads, as created by arrow().}

\item{arrow_size}{The size of the arrow relative to the line size.}

\item{shadow}{Shadow specification as created by \code{\link[=shadow]{shadow()}}}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{data}{A function created by \code{\link[=construct_get_segment_info]{construct_get_segment_info()}}.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}
}
\description{
Plotting individual segments of the trajectory
}
