% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_velocity_.R
\name{geom_velocity_arrow}
\alias{geom_velocity_arrow}
\title{Plotting velocity}
\usage{
geom_velocity_arrow(mapping = NULL, stat = stat_velocity_grid(),
  arrow = ggplot2::arrow(length = ggplot2::unit(0.2, "cm")), ...,
  data = construct_get_velocity_info(stat), show.legend = NA)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{stat}{Where to place the arrows, such as for every cell (\code{\link[=stat_velocity_cells]{stat_velocity_cells()}}) or using a grid (\code{\link[=stat_velocity_grid]{stat_velocity_grid()}})}

\item{arrow}{specification for arrow heads, as created by arrow().}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{data}{A function created by \code{\link[=construct_get_velocity_info]{construct_get_velocity_info()}}}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}
}
\description{
Plotting velocity
}
