% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{milestone_palette}
\alias{milestone_palette}
\alias{get_milestone_palette_names}
\title{Get the names of valid color palettes}
\usage{
milestone_palette(n, name = "auto")

get_milestone_palette_names()
}
\arguments{
\item{n}{The number of colours to be in the palette.}

\item{name}{The name of the palette. Must be one of \code{"cubeHelix"}, \code{"Set3"}, or \code{"rainbow"}. \code{auto} will choose a palette depending on the n}
}
\description{
Get the names of valid color palettes
}
\keyword{plot_helpers}
