% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_dendro.R
\name{layout_dendro}
\alias{layout_dendro}
\title{Dendrogram layout of a trajectory}
\usage{
layout_dendro(trajectory, diag_offset = 0.05)
}
\arguments{
\item{trajectory}{The trajectory as created by \code{\link[=infer_trajectory]{infer_trajectory()}} or \code{\link[=add_trajectory]{add_trajectory()}}}

\item{diag_offset}{The x-offset (percentage of the edge lenghts) between milestones}
}
\description{
Dendrogram layout of a trajectory
}
\keyword{plot_trajectory}
