% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearise_cells.R
\name{linearise_cells}
\alias{linearise_cells}
\title{Linearise a trajectory}
\usage{
linearise_cells(milestone_network, progressions, margin = 0.05,
  one_edge = FALSE)
}
\arguments{
\item{milestone_network}{The milestone network}

\item{progressions}{The progressions}

\item{margin}{The margin to add}

\item{one_edge}{If TRUE, assigns each cell to one edge only}
}
\description{
Linearise a trajectory
}
