% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_linearise_cells.R
\name{linearise_trajectory}
\alias{linearise_trajectory}
\title{Linearise a trajectory}
\usage{
linearise_trajectory(trajectory, margin = 0.05,
  no_margin_between_linear = TRUE, one_edge = TRUE,
  equal_cell_width = FALSE)
}
\arguments{
\item{trajectory}{The trajectory}

\item{margin}{The margin to add}

\item{no_margin_between_linear}{Whether to remove the margin if a milestone does not split or converge}

\item{one_edge}{Whether to assign each cell to an edge only once. Only relevant in case of divergence regions}

\item{equal_cell_width}{Whether to give each cell an equal width. Useful when plotting heatmaps.}
}
\description{
Linearise a trajectory
}
