% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{order_cells}
\alias{order_cells}
\title{order the cells according to their progression, assign tented cells to the highest percentage}
\usage{
order_cells(milestone_network, progressions)
}
\description{
order the cells according to their progression, assign tented cells to the highest percentage
}
