% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{plot_combined}
\alias{plot_combined}
\title{Plot a dimensionality reduced trajectory}
\usage{
plot_combined(original_object, new_object, transition_size, cell_size,
  milestone_size, arrow_length, plot_label, plot_milestones)
}
\arguments{
\item{original_object}{An object generated by \code{wrap_data} or \code{dimred_trajectory}.}

\item{new_object}{An object generated by \code{wrap_data} or \code{dimred_trajectory}.}

\item{transition_size}{The size of the transition lines between milestones.}

\item{cell_size}{The size of cells.}

\item{milestone_size}{The size of milestones.}

\item{arrow_length}{length of the arrow.}

\item{plot_label}{What to label. Must be one of \code{"leaves"}, \code{"all"}, or \code{"none"}.}

\item{plot_milestones}{Whether to plot the milestones.}
}
\description{
Plot a dimensionality reduced trajectory
}
