% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_connections.R
\name{plot_connections}
\alias{plot_connections}
\title{Plot connections}
\usage{
plot_connections(milestone_network, orientation = 1, plotdata = NULL)
}
\arguments{
\item{milestone_network}{The milestone network}

\item{orientation}{?? TODO: Zouter/wouters}

\item{plotdata}{?? TODO: Zouter/wouters}
}
\description{
Plot connections
}
