% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{plot_default}
\alias{plot_default}
\title{Plot a dimensionality reduced trajectory}
\usage{
plot_default(object, insert_phantom_edges = TRUE)
}
\arguments{
\item{object}{output as generated by \code{\link{dimred_trajectory}}, \code{\link{wrap_ti_task_data}} or \code{\link[dynmethods]{wrap_prediction_model}}}

\item{insert_phantom_edges}{attempt to plot bifurcating edges correctly automatically}
}
\description{
Plot a dimensionality reduced trajectory
}
