% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dendro.R
\name{plot_dendro}
\alias{plot_dendro}
\title{Plot a tree trajectory as a dendrogram}
\usage{
plot_dendro(task, grouping = NULL, groups = NULL)
}
\arguments{
\item{task}{The trajectory}

\item{diag_offset}{The x-offset (percentage of the edge lenghts) between milestones}
}
\description{
Plot a tree trajectory as a dendrogram
}
