% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dimred.R
\name{plot_dimred}
\alias{plot_dimred}
\title{Plot trajectory on dimensionality reduction}
\usage{
plot_dimred(task, color_cells = c("auto", "invisible", "positions",
  "grouping", "gene", "milestone", "pseudotime"), grouping_assignment = NULL,
  groups = NULL, gene_oi = NULL, color_milestones = c("auto", "given"),
  milestones = NULL, pseudotime = NULL, expression_source = "expression",
  cell_positions,
  plot_milestone_network = dynwrap::is_wrapper_with_trajectory(task),
  dimred_method = dimred_pca)
}
\arguments{
\item{color_cells}{How to color the cells}

\item{grouping_assignment}{Tibble containing the assignment of cells to groups of cells}

\item{groups}{Tibble containing information of the cell groups}

\item{gene_oi}{Gene to plot expression}

\item{color_milestones}{How to color the cells}

\item{milestones}{Tibble containing the `milestone_id` and a `color` for each milestone}

\item{pseudotime}{The pseudotime}

\item{expression_source}{Source of the expression}
}
\description{
Plot trajectory on dimensionality reduction
}
