% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_edge_flips.R
\name{plot_edge_flips}
\alias{plot_edge_flips}
\title{Plotting edge flips}
\usage{
plot_edge_flips(oldadj, newadj)
}
\arguments{
\item{oldadj}{Old adjacency matrix}

\item{newadj}{New adjancency matrix}
}
\description{
Utility plotting function for analysing the edge-flip metric from dyneval.
}
\keyword{compare_trajectory}
