% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes.R
\name{plot_genes}
\alias{plot_genes}
\title{Plotting the genes}
\usage{
plot_genes(task, genes = sample(colnames(task$counts), min(c(20,
  ncol(task$counts)))), margin = 0.05)
}
\arguments{
\item{task}{The task}

\item{genes}{The genes}

\item{margin}{The margin to add}
}
\description{
Plotting the genes
}
