% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot the task as a heatmap}
\usage{
plot_heatmap(task, genes_oi = colnames(task$counts)[1:20],
  clust = hclust(as.dist(correlation_distance(t(task$counts[, genes_oi]))),
  method = "ward.D2"), margin = 0.02)
}
\arguments{
\item{task}{The task}

\item{genes_oi}{Genes to plot}

\item{clust}{The clustering of the genes as a `clust` object}

\item{margin}{The margin to add}
}
\description{
Plot the task as a heatmap
}
