% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_strip.R
\name{plot_strip}
\alias{plot_strip}
\title{Plot strip}
\usage{
plot_strip(task1, task2, margin = 0.05, reorder = TRUE)
}
\arguments{
\item{task1}{The first task}

\item{task2}{The second task}

\item{margin}{The margin to add}

\item{reorder}{?? TODO: Zouter/wouters}
}
\description{
Plot strip
}
