% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_strip_connections.R
\name{plot_strip_connections}
\alias{plot_strip_connections}
\title{Plot strip connections}
\usage{
plot_strip_connections(task1, task2, reorder = TRUE, margin = 0.05,
  reorder_second_by = "mapping")
}
\arguments{
\item{task1}{The first task}

\item{task2}{The second task}

\item{reorder}{?? TODO: Zouter/wouters}

\item{margin}{The margin to add}

\item{reorder_second_by}{?? TODO: Zouter/wouters}
}
\description{
Plot strip connections
}
