% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_trajectory_segments.R
\name{position_trajectory_arrows_middle}
\alias{position_trajectory_arrows_middle}
\alias{position_trajectory_arrows_boundaries}
\title{Plot arrow in middle}
\usage{
position_trajectory_arrows_middle()

position_trajectory_arrows_boundaries(quantile = 0.3)
}
\arguments{
\item{quantile}{The quantile at which to plot an arrow}
}
\description{
Plot arrow in middle

Plot arrow at certain quantiles, be default at the beginning and end
}
