% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dimred.R
\name{project_waypoints}
\alias{project_waypoints}
\title{Project the waypoints}
\usage{
project_waypoints(trajectory, cell_positions, edge_positions = NULL,
  waypoints = dynwrap::select_waypoints(trajectory),
  trajectory_projection_sd = sum(trajectory$milestone_network$length) *
  0.05, color_trajectory = "none")
}
\arguments{
\item{trajectory}{The trajectory}

\item{cell_positions}{The positions of the cells}

\item{waypoints}{The waypoints to use for projecting, as generated by [dynwrap::select_waypoints()]}

\item{trajectory_projection_sd}{The standard deviation of the gaussian kernel}

\item{color_trajectory}{How to color the trajectory, can be "nearest" for coloring to nearest cell, or "none"}
}
\description{
Project the waypoints
}
