% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_clean.R
\name{theme_clean}
\alias{theme_clean}
\title{We like our plots clean}
\usage{
theme_clean()
}
\value{
A ggplot2 theme.
}
\description{
We like our plots clean
}
\examples{
data(example_bifurcating)
g <- plot_dimred(example_bifurcating)
g + theme_clean()
}
\keyword{plot_helpers}
