% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cell_coloring.R
\name{add_cell_coloring}
\alias{add_cell_coloring}
\title{Add coloring}
\usage{
add_cell_coloring(cell_positions, color_cells = c("auto", "invisible",
  "positions", "grouping", "gene", "milestone", "pseudotime"), task,
  grouping_assignment = NULL, groups = NULL, gene_oi = NULL,
  expression_source = "expression", pseudotime = NULL,
  color_milestones = c("auto", "given"), milestones = NULL)
}
\arguments{
\item{color_cells}{How to color the cells}

\item{task}{The task}

\item{grouping_assignment}{Tibble containing the assignment of cells to groups of cells}

\item{groups}{Tibble containing information of the cell groups}

\item{gene_oi}{Gene to plot expression}

\item{expression_source}{Source of the gene expression, defaults to `expression`}

\item{pseudotime}{The pseudotime}

\item{color_milestones}{How to color the cells}

\item{milestones}{Tibble containing the `milestone_id` and a `color` for each milestone}

\item{cell_position}{The positions of the cells}
}
\description{
Add coloring
}
