% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cell_coloring.R
\name{add_cell_coloring}
\alias{add_cell_coloring}
\title{Add colouring to a set of cells.}
\usage{
add_cell_coloring(
  cell_positions,
  color_cells = c("auto", "none", "grouping", "feature", "milestone", "pseudotime"),
  trajectory,
  grouping = NULL,
  groups = NULL,
  feature_oi = NULL,
  expression_source = "expression",
  pseudotime = NULL,
  color_milestones = c("auto", "given", "cubeHelix", "Set3", "rainbow"),
  milestones = NULL,
  milestone_percentages = NULL
)
}
\arguments{
\item{cell_positions}{The positions of the cells, represented by a tibble.
Must contain column `cell_id` (character) and may contain columns `from`,
`to`, `pseudotime`, depending on the value of `color_cells`.}

\item{color_cells}{How to color the cells}

\item{trajectory}{The trajectory}

\item{grouping}{The grouping of the cells}

\item{groups}{Tibble containing information of the cell groups}

\item{feature_oi}{feature to plot expression}

\item{expression_source}{Source of the feature expression, defaults to `expression`}

\item{pseudotime}{The pseudotime}

\item{color_milestones}{Which palette to use for colouring the milestones
* `auto`: Determine colours automatically. If `color` is already specified in
  milestones tibble, this will be used. Otherwise, the colour scheme is determined by
  `milestone_palette_list$auto`.
* `given`: The `milestones` object already contains a column `color`.
* `cubeHelix`: Use the `rje::cubeHelix()` palette.
* `Set3`: Use the `RColorBrewer::brewer.pal(name = "Set3")` palette.
* `rainbow`: Use the `grDevices::rainbow()` palette.}

\item{milestones}{Tibble containing the column `milestone_id` (character).
If `color_milestones` is set to `"given"`, this tibble should also contain a column
`color` (character), containing colour hex codes (e.g. `"#123456"`).}

\item{milestone_percentages}{The milestone percentages}
}
\description{
Add colouring to a set of cells.
}
