% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coloring.R
\name{add_cell_coloring}
\alias{add_cell_coloring}
\title{Add coloring}
\usage{
add_cell_coloring(
  cell_positions,
  color_cells = c("auto", "none", "grouping", "feature", "milestone", "pseudotime"),
  dataset,
  trajectory = dataset,
  grouping = NULL,
  groups = NULL,
  feature_oi = NULL,
  expression_source = "expression",
  pseudotime = NULL,
  color_milestones = c("auto", "given", "cubeHelix", "Set3", "rainbow"),
  milestones = NULL,
  milestone_percentages = NULL
)
}
\arguments{
\item{cell_positions}{The positions of the cells}

\item{color_cells}{How to color the cells}

\item{trajectory}{The trajectory}

\item{grouping}{The grouping of the cells}

\item{groups}{Tibble containing information of the cell groups}

\item{feature_oi}{feature to plot expression}

\item{expression_source}{Source of the feature expression, defaults to `expression`}

\item{pseudotime}{The pseudotime}

\item{color_milestones}{How to color the cells}

\item{milestones}{Tibble containing the `milestone_id` and a `color` for each milestone}

\item{milestone_percentages}{The milestone percentages}
}
\description{
Add coloring
}
