% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory_dimred.R
\name{dimred_trajectory}
\alias{dimred_trajectory}
\title{Perform dimensionality reduction on a trajectory and the respective samples in order to plot it}
\usage{
dimred_trajectory(traj_object, insert_phantom_edges = TRUE,
  color_by = c("Set3", "rainbow"))
}
\arguments{
\item{traj_object}{A trajectory as generated by
\code{\link[dynwrap]{wrap_data}}}

\item{insert_phantom_edges}{attempt to plot bifurcating edges correctly automatically}

\item{color_by}{How to color the edges, either using Set3 or using rainbow}
}
\description{
Perform dimensionality reduction on a trajectory and the respective samples in order to plot it
}
