% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{dynplot}
\alias{dynplot}
\title{Plot all the trajectories}
\description{
Create ggplot2 plots of single-cell trajectories. This can be
the trajectory topology itself, projected onto a dimensionality reduction,
using a heatmap, or a comparison between two trajectories.
}
