% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R, R/package.R
\docType{package}
\name{dynplot}
\alias{dynplot}
\alias{dynplot-package}
\title{Plot all the trajectories}
\usage{
dynplot(dataset)
}
\description{
Plot all the trajectories
}
\examples{
toy <- dyntoy::generate_dataset()
dataset <- dynwrap::infer_trajectory(toy, dynmethods::ti_scorpius())

}
