% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{dynplot}
\alias{dynplot}
\title{dynplot: Plotting Single-Cell Trajectories}
\description{
Visualise a single-cell trajectory as a graph or dendrogram,
as a dimensionality reduction or heatmap of the expression data,
or a comparison between two trajectories as a pairwise scatterplot
or dimensionality reduction projection.
}
