% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone_palette.R
\name{milestone_palette}
\alias{milestone_palette}
\alias{get_milestone_palette_names}
\title{Get the names of valid color palettes}
\usage{
milestone_palette(name, n)

get_milestone_palette_names()
}
\arguments{
\item{name}{The name of the palette. Must be one of \code{"cubeHelix"}, \code{"Set3"}, or \code{"rainbow"}.}

\item{n}{The number of colours to be in the palette.}
}
\value{
The names of supported palettes.
}
\description{
Get the names of valid color palettes
}
\examples{
get_milestone_palette_names()
}
\keyword{plot_helpers}
