% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_connections.R
\name{make_connection_plotdata}
\alias{make_connection_plotdata}
\title{Make connection plotdata}
\usage{
make_connection_plotdata(milestone_network, orientation = 1, margin = 0.05)
}
\arguments{
\item{milestone_network}{The milestone network}

\item{orientation}{The orientation of the onedim between milestones}

\item{margin}{Relative distance between milestones}
}
\description{
Make connection plotdata
}
