% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{plot_combined}
\alias{plot_combined}
\title{Plot a dimensionality reduced trajectory}
\usage{
plot_combined(original_object, new_object, insert_phantom_edges = TRUE)
}
\arguments{
\item{original_object}{output as generated by \code{\link{dimred_trajectory}}, \code{\link{wrap_ti_task_data}} or \code{\link{wrap_ti_prediction}}}

\item{new_object}{output as generated by \code{\link{dimred_trajectory}}, \code{\link{wrap_ti_task_data}} or \code{\link{wrap_ti_prediction}}}

\item{insert_phantom_edges}{attempt to plot bifurcating edges correctly automatically}
}
\description{
Plot a dimensionality reduced trajectory
}
