% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_default.R
\name{plot_default}
\alias{plot_default}
\title{Plot a dimensionality reduced trajectory}
\usage{
plot_default(object, colour_cells = "milestone", colour_milestones = "auto",
  transition_size = 8, cell_size = 3, milestone_size = 5,
  arrow_length = grid::unit(1, "cm"), plot_label = c("leaves", "all",
  "none"), plot_milestones = TRUE)
}
\arguments{
\item{object}{An object generated by \code{wrap_data} or \code{dimred_trajectory}.}

\item{colour_cells}{How to colour the cells. Must be equal to \code{"milestone"}
or a custom character vector of colours, one for each cell.}

\item{colour_milestones}{How to color the milestones.
Must be one of \code{"auto"} (default), \code{"Set3"}, \code{"rainbow"}, \code{"cubeHelix"},
or a custom character vector of colours, one for each milestone.}

\item{transition_size}{The size of the transition lines between milestones.}

\item{cell_size}{The size of cells.}

\item{milestone_size}{The size of milestones.}

\item{arrow_length}{length of the arrow.}

\item{plot_label}{What to label. Must be one of \code{"leaves"}, \code{"all"}, or \code{"none"}.}

\item{plot_milestones}{Whether to plot the milestones.}
}
\description{
Plot a dimensionality reduced trajectory
}
