% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{plot_default}
\alias{plot_default}
\title{Plot a dimensionality reduced trajectory}
\usage{
plot_default(object, insert_phantom_edges = TRUE, line_size = 8,
  arrow_length = unit(1, "cm"), label = c("leaves", "all", "none"),
  plot_milestones = TRUE)
}
\arguments{
\item{object}{output as generated by \code{\link{dimred_trajectory}} or \code{\link[dynwrap]{wrap_data}}}

\item{insert_phantom_edges}{attempt to plot bifurcating edges correctly automatically}

\item{line_size}{size of the line}

\item{arrow_length}{length of the arrow}

\item{label}{what to label: the leaves, all nodes or none}

\item{plot_milestones}{whether to plot the milestones}
}
\description{
Plot a dimensionality reduced trajectory
}
