% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dendro.R
\name{plot_dendro}
\alias{plot_dendro}
\title{Plot a tree trajectory as a dendrogram}
\usage{
plot_dendro(task, color_cells = c("auto", "invisible", "black", "positions",
  "grouping", "gene"), grouping_assignment = NULL, groups = NULL,
  gene_oi = NULL, expression_source = "expression", diag_offset = 0.05)
}
\arguments{
\item{task}{The trajectory}

\item{color_cells}{How to color the cells, one of auto, none, grouping}

\item{grouping_assignment}{Tibble containing the assignment of cells to groups of cells}

\item{groups}{Tibble containing information of the cell groups}

\item{gene_oi}{Gene to plot expression}

\item{expression_source}{Source of the gene expression, defaults to `expression`}

\item{diag_offset}{The x-offset (percentage of the edge lenghts) between milestones}
}
\description{
Plot a tree trajectory as a dendrogram
}
