% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dendro.R
\name{plot_dendro}
\alias{plot_dendro}
\title{Plot a trajectory as a dendrogram}
\usage{
plot_dendro(
  trajectory,
  color_cells = c("auto", "none", "grouping", "feature", "milestone", "pseudotime"),
  grouping = NULL,
  groups = NULL,
  feature_oi = NULL,
  expression_source = "expression",
  pseudotime = NULL,
  color_milestones = c("auto", "given", "cubeHelix", "Set3", "rainbow"),
  milestones = NULL,
  milestone_percentages = NULL,
  alpha_cells = 1,
  size_cells = 2.5,
  border_radius_percentage = 0.1,
  diag_offset = 0.05,
  y_offset = 0.2,
  arrow = grid::arrow(type = "closed")
)
}
\arguments{
\item{trajectory}{The trajectory}

\item{color_cells}{How to color the cells}

\item{grouping}{The grouping of the cells}

\item{groups}{Tibble containing information of the cell groups}

\item{feature_oi}{feature to plot expression}

\item{expression_source}{Source of the feature expression, defaults to `expression`}

\item{pseudotime}{The pseudotime}

\item{color_milestones}{How to color the cells}

\item{milestones}{Tibble containing the `milestone_id` and a `color` for each milestone}

\item{milestone_percentages}{The milestone percentages}

\item{alpha_cells}{The alpha of the cells}

\item{size_cells}{The size of the cells}

\item{border_radius_percentage}{The fraction of the radius that is used for the border}

\item{diag_offset}{The x-offset (percentage of the edge lenghts) between milestones}

\item{y_offset}{The size of the quasirandom cell spreading in the y-axis}

\item{arrow}{The type and size of arrow in case of directed trajectories. Set to NULL to remove arrow altogether.}
}
\value{
A dendrogram ggplot of the trajectory.
}
\description{
Plot a trajectory as a dendrogram
}
\examples{
data(example_tree)
plot_dendro(example_tree)
plot_dendro(example_tree, color_cells = "pseudotime")
plot_dendro(
  example_tree,
  color_cells = "grouping",
  grouping = dynwrap::group_onto_nearest_milestones(example_tree)
)
}
\keyword{plot_trajectory}
