% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dendro.R
\name{plot_dendro}
\alias{plot_dendro}
\title{Plot a tree trajectory as a dendrogram}
\usage{
plot_dendro(traj, color_cells = c("auto", "none", "grouping", "feature",
  "milestone", "pseudotime"), grouping = NULL, groups = NULL,
  feature_oi = NULL, expression_source = "expression",
  pseudotime = NULL, color_milestones = c("auto", "given", "cubeHelix",
  "Set3", "rainbow"), milestones = NULL, milestone_percentages = NULL,
  diag_offset = 0.05, y_offset = 0.2)
}
\arguments{
\item{traj}{The trajectory}

\item{color_cells}{How to color the cells}

\item{grouping}{The grouping of the cells}

\item{groups}{Tibble containing information of the cell groups}

\item{feature_oi}{feature to plot expression}

\item{expression_source}{Source of the feature expression, defaults to `expression`}

\item{pseudotime}{The pseudotime}

\item{color_milestones}{How to color the cells}

\item{milestones}{Tibble containing the `milestone_id` and a `color` for each milestone}

\item{milestone_percentages}{The milestone percentages}

\item{diag_offset}{The x-offset (percentage of the edge lenghts) between milestones}

\item{y_offset}{The size of the quasirandom cell spreading in the y-axis}
}
\description{
Plot a tree trajectory as a dendrogram
}
\keyword{plot_trajectory}
