% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dimred.R
\name{plot_dimred}
\alias{plot_dimred}
\title{Plot a trajectory on dimensionality reduction}
\usage{
plot_dimred(
  dataset,
  trajectory = dataset,
  color_cells = c("auto", "none", "grey", "grouping", "feature", "milestone",
    "pseudotime"),
  dimred = ifelse(dynwrap::is_wrapper_with_dimred(trajectory), NA,
    ifelse(length(trajectory$cell_ids) > 500, dimred_pca, dimred_mds)),
  size_cells = log10(min(10000, length(dataset$cell_ids)))/4,
  alpha_cells = 1,
  hex_cells = ifelse(length(dataset$cell_ids) > 10000, 100, FALSE),
  plot_trajectory = "auto",
  plot_milestones = "auto",
  color_milestones = c("auto", "given", "cubeHelix", "Set3", "rainbow"),
  color_trajectory = c("none"),
  shadow_trajectory = dynplot2::shadow_defaults(size = 1.5),
  shadow_milestones = shadow_trajectory,
  size_trajectory = 2,
  size_trajectory_arrow = 0.5,
  size_milestones = size_trajectory * 3,
  grouping = NULL,
  groups = NULL,
  feature_oi = NULL,
  milestones = NULL,
  milestone_percentages = NULL,
  pseudotime = NULL,
  expression_source = "expression",
  plot_contour = "none",
  contour_relative_bandwidth = 0.2,
  contour_density_cutoff = 0.2,
  plot_velocity = case_when(!is.null(dataset$velocity) ~ "stream", TRUE ~ "none")
)
}
\arguments{
\item{hex_cells}{The number of hexes to use, to avoid overplotting points. Default is FALSE if number of cells <= 10000
#' @include layer_milestones.R layer_trajectory.R layer_cells.R layer_velocity.R layer_contour.R}
}
\description{
Plot a trajectory on dimensionality reduction
}
