% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dimred.R
\name{plot_dimred}
\alias{plot_dimred}
\title{Plot trajectory on dimensionality reduction}
\usage{
plot_dimred(traj, color_cells = c("auto", "none", "grouping", "feature",
  "milestone", "pseudotime"), color_density = c("none", "grouping",
  "feature"), grouping_assignment = NULL, groups = NULL,
  feature_oi = NULL, color_milestones = c("auto", "given"),
  milestones = NULL, milestone_percentages = NULL, pseudotime = NULL,
  expression_source = "expression",
  plot_milestone_network = dynwrap::is_wrapper_with_trajectory(traj),
  label_milestones = dynwrap::is_wrapper_with_milestone_labelling(traj),
  dimred = ifelse(dynwrap::is_wrapper_with_dimred(traj), NA,
  ifelse(length(traj$cell_ids) > 500, dimred_pca, dimred_mds)), padding = 0.1,
  nbins = 1000, bw = 0.2, density_cutoff = 0.3,
  density_cutoff_label = density_cutoff/10)
}
\arguments{
\item{traj}{The trajectory}

\item{color_cells}{How to color the cells}

\item{color_density}{How to color density, can be "none", "grouping", or "feature"}

\item{grouping_assignment}{Tibble containing the assignment of cells to groups of cells}

\item{groups}{Tibble containing information of the cell groups}

\item{feature_oi}{feature to plot expression}

\item{color_milestones}{How to color the cells}

\item{milestones}{Tibble containing the `milestone_id` and a `color` for each milestone}

\item{milestone_percentages}{The milestone percentages}

\item{pseudotime}{The pseudotime}

\item{expression_source}{Source of the expression}

\item{plot_milestone_network}{Whether to plot the milestone network}

\item{label_milestones}{How to label the milestones. Can be TRUE (in which case the labels within the trajectory will be used), "all" (in which case both given labels as milestone_ids will be used), a named character vector, or FALSE}

\item{dimred}{The dimensionality reduction matrix (with cell_ids as rownames) or function which will run the dimensionality reduction}

\item{padding}{The padding in the edges to the plot, relative to the size of the plot}

\item{nbins}{Number of bins for calculating the density}

\item{bw}{Bandwidth, relative to the size of the plot}

\item{density_cutoff}{Cutoff for density, the lower the larger the areas}

\item{density_cutoff_label}{Cutoff for density for labelling, the lower the further way from cells}
}
\description{
Plot trajectory on dimensionality reduction
}
