% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dimred.R
\name{plot_dimred}
\alias{plot_dimred}
\title{Plot a trajectory on dimensionality reduction}
\usage{
plot_dimred(
  trajectory,
  color_cells = c("auto", "none", "grouping", "feature", "milestone", "pseudotime"),
  dimred = dyndimred::dimred_landmark_mds,
  plot_trajectory = dynwrap::is_wrapper_with_trajectory(trajectory) &&
    !plot_milestone_network,
  plot_milestone_network = FALSE,
  label_milestones = dynwrap::is_wrapper_with_milestone_labelling(trajectory),
  alpha_cells = 1,
  size_cells = 2.5,
  border_radius_percentage = 0.1,
  size_milestones = 6,
  size_transitions = 1,
  hex_cells = ifelse(length(trajectory$cell_ids) > 10000, 100, FALSE),
  grouping = NULL,
  groups = NULL,
  feature_oi = NULL,
  color_milestones = c("auto", "given", "cubeHelix", "Set3", "rainbow"),
  milestones = NULL,
  milestone_percentages = NULL,
  pseudotime = NULL,
  expression_source = "expression",
  arrow = grid::arrow(type = "closed", length = unit(0.1, "inches")),
  color_density = c("none", "grouping", "feature"),
  padding = 0.1,
  nbins = 1000,
  bw = 0.2,
  density_cutoff = 0.3,
  density_cutoff_label = density_cutoff/10,
  waypoints = dynwrap::select_waypoints(trajectory),
  trajectory_projection_sd = sum(trajectory$milestone_network$length) * 0.05,
  color_trajectory = "none"
)
}
\arguments{
\item{trajectory}{The trajectory}

\item{color_cells}{How to color the cells}

\item{dimred}{Can be
\itemize{
\item A function which will perform the dimensionality reduction, see \code{\link[dyndimred:dimred]{dyndimred::list_dimred_methods()}}
\item A matrix with the dimensionality reduction, with cells in rows and dimensions (\emph{comp_1}, \emph{comp_2}, ...) in columns
}}

\item{plot_trajectory}{Whether to plot the projected trajectory on the dimensionality reduction}

\item{plot_milestone_network}{Whether to plot the projected milestone network on the dimensionality reduction}

\item{label_milestones}{How to label the milestones. Can be TRUE (in which case the labels within the trajectory will be used), "all" (in which case both given labels and milestone_ids will be used), a named character vector, or FALSE}

\item{alpha_cells}{The alpha of the cells}

\item{size_cells}{The size of the cells}

\item{border_radius_percentage}{The fraction of the radius that is used for the border}

\item{size_milestones}{The size of the milestones}

\item{size_transitions}{The size of the trajectory segments}

\item{hex_cells}{The number of hexes to use, to avoid overplotting points. Default is FALSE if number of cells <= 10000.}

\item{grouping}{The grouping of the cells}

\item{groups}{Tibble containing information of the cell groups}

\item{feature_oi}{feature to plot expression}

\item{color_milestones}{How to color the cells}

\item{milestones}{Tibble containing the `milestone_id` and a `color` for each milestone}

\item{milestone_percentages}{The milestone percentages}

\item{pseudotime}{The pseudotime}

\item{expression_source}{Source of the expression}

\item{arrow}{The type and size of arrow in case of directed trajectories. Set to NULL to remove arrow altogether.}

\item{color_density}{How to color density, can be "none", "grouping", or "feature"}

\item{padding}{The padding in the edges to the plot, relative to the size of the plot}

\item{nbins}{Number of bins for calculating the density}

\item{bw}{Bandwidth, relative to the size of the plot}

\item{density_cutoff}{Cutoff for density, the lower the larger the areas}

\item{density_cutoff_label}{Cutoff for density for labelling, the lower the further way from cells}

\item{waypoints}{The waypoints to use for projecting, as generated by [dynwrap::select_waypoints()]}

\item{trajectory_projection_sd}{The standard deviation of the Gaussian kernel to be used for projecting the trajectory.
This is in the order of magnitude as the lengths of the milestone_network.
The lower, the more closely the trajectory will follow the cells.}

\item{color_trajectory}{How to color the trajectory, can be "nearest" for coloring to nearest cell, or "none"}
}
\description{
Plot a trajectory on dimensionality reduction
}
\examples{
\donttest{
data(example_bifurcating)
plot_dimred(example_bifurcating)
plot_dimred(example_bifurcating, dimred = dyndimred::dimred_umap)

dimred <- dyndimred::dimred_dm_destiny(example_bifurcating$expression)
plot_dimred(example_bifurcating, dimred = dimred)

plot_dimred(example_bifurcating, color_cells = "pseudotime")
plot_dimred(
  example_bifurcating,
  color_density = "grouping",
  grouping = dynwrap::group_onto_nearest_milestones(example_bifurcating)
)
}

}
\keyword{plot_trajectory}
