% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dimred.R
\name{plot_dimred}
\alias{plot_dimred}
\title{Plot trajectory on dimensionality reduction}
\usage{
plot_dimred(task, color_cells = c("auto", "invisible", "positions",
  "grouping", "gene", "milestone", "pseudotime"), grouping_assignment = NULL,
  groups = NULL, gene_oi = NULL, color_milestones = c("auto", "given",
  get_milestone_palette_names()), milestones = NULL,
  milestone_percentages = NULL, pseudotime = NULL,
  expression_source = "expression", cell_positions,
  plot_milestone_network = dynwrap::is_wrapper_with_trajectory(task),
  plot_milestone_labels = TRUE, dimred_method = ifelse(length(task$cell_ids)
  > 500, dimred_pca, dimred_mds))
}
\arguments{
\item{task}{The task}

\item{color_cells}{How to color the cells}

\item{grouping_assignment}{Tibble containing the assignment of cells to groups of cells}

\item{groups}{Tibble containing information of the cell groups}

\item{gene_oi}{Gene to plot expression}

\item{color_milestones}{How to color the cells}

\item{milestones}{Tibble containing the `milestone_id` and a `color` for each milestone}

\item{milestone_percentages}{The milestone percentages}

\item{pseudotime}{The pseudotime}

\item{expression_source}{Source of the expression}

\item{cell_positions}{The positions of the cells}

\item{plot_milestone_network}{Whether to plot the milestone network}

\item{plot_milestone_labels}{Whether to overlay the milestone labels over the network}

\item{dimred_method}{The dimred method or the dimensionality reduction (a dataframe containing at least Comp1, Comp2 and cell_id)}
}
\description{
Plot trajectory on dimensionality reduction
}
