% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_features.R
\name{plot_features}
\alias{plot_features}
\title{Plotting a set of features in a line plot}
\usage{
plot_features(traj, expression_source = "expression", features_oi = 2,
  margin = 0.02, cell_feature_importances = NULL, scale = TRUE)
}
\arguments{
\item{features_oi}{The features of interest}

\item{margin}{The margin to add}
}
\description{
Plotting a set of features in a line plot
}
