% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_graph}
\alias{plot_graph}
\alias{plot_default}
\title{Plot a trajectory and cellular positions as a graph}
\usage{
plot_graph(
  trajectory,
  color_cells = c("auto", "none", "grouping", "feature", "milestone", "pseudotime"),
  color_milestones = c("auto", "given", "cubeHelix", "Set3", "rainbow"),
  grouping = NULL,
  groups = NULL,
  feature_oi = NULL,
  pseudotime = NULL,
  expression_source = "expression",
  milestones = NULL,
  milestone_percentages = NULL,
  size_trajectory = 3,
  size_milestones = 8,
  alpha_cells = 1,
  size_cells = 2.5,
  border_radius_percentage = 0.1,
  arrow = grid::arrow(length = grid::unit(1, "cm"), type = "closed"),
  label_milestones = dynwrap::is_wrapper_with_milestone_labelling(trajectory),
  plot_milestones = FALSE,
  adjust_weights = FALSE
)
}
\arguments{
\item{trajectory}{The trajectory as created by \code{\link[dynwrap:infer_trajectories]{infer_trajectory()}} or \code{\link[dynwrap:add_trajectory]{add_trajectory()}}}

\item{color_cells}{How to color the cells}

\item{color_milestones}{How to color the cells}

\item{grouping}{The grouping of the cells}

\item{groups}{Tibble containing information of the cell groups}

\item{feature_oi}{feature to plot expression}

\item{pseudotime}{The pseudotime}

\item{expression_source}{Source of the feature expression, defaults to `expression`}

\item{milestones}{Tibble containing the `milestone_id` and a `color` for each milestone}

\item{milestone_percentages}{The milestone percentages}

\item{size_trajectory}{The size of the transition lines between milestones.}

\item{size_milestones}{The size of milestones.}

\item{alpha_cells}{The alpha of the cells.}

\item{size_cells}{The size of the cells.}

\item{border_radius_percentage}{The fraction of the radius that is used for the border.}

\item{arrow}{The type and size of arrow in case of directed trajectories. Set to NULL to remove arrow altogether.}

\item{label_milestones}{How to label the milestones. Can be TRUE (in which case the labels within the trajectory will be used), "all" (in which case both given labels and milestone_ids will be used), a named character vector, or FALSE}

\item{plot_milestones}{Whether to plot the milestones.}

\item{adjust_weights}{Whether or not to rescale the milestone network weights}
}
\value{
A graph ggplot of a trajectory.
}
\description{
Plot a trajectory and cellular positions as a graph
}
\examples{
data(example_disconnected)
plot_graph(example_disconnected)
plot_graph(example_disconnected, color_cells = "pseudotime")
plot_graph(
  example_disconnected,
  color_cells = "grouping",
  grouping = dynwrap::group_onto_nearest_milestones(example_disconnected)
)

data(example_tree)
plot_graph(example_tree)
}
\keyword{plot_trajectory}
