% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap2.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot the expression across a trajectory in a heatmap}
\usage{
plot_heatmap(
  dataset,
  trajectory,
  expression_source = dataset,
  features_oi = 20,
  clust = "ward.D2",
  color_cells = NULL,
  milestones,
  milestone_percentages = trajectory$milestone_percentages,
  grouping = NULL,
  groups = NULL,
  scale = dynutils::scale_quantile,
  column_gap = unit(1, "mm"),
  plot_milestone_percentages = c("top", "bottom", "none")
)
}
\arguments{
\item{features_oi}{The features of interest, either the number of features or a vector giving the names of the different features}

\item{clust}{The method to cluster the features, or a hclust object}

\item{scale}{Whether to rescale the expression, can be a function or boolean}
}
\description{
Plot the expression across a trajectory in a heatmap
}
\examples{
data(example_linear)
plot_heatmap(example_linear)

data(example_bifurcating)
plot_heatmap(example_bifurcating)

}
\keyword{plot_trajectory}
