% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot the expression across a trajectory in a heatmap}
\usage{
plot_heatmap(
  dataset,
  trajectory = dataset,
  expression_source = dataset,
  features_oi,
  feature_info = dataset$feature_info,
  feature_annotation = NULL,
  row_gap = unit(3, "mm"),
  clust = "ward.D2",
  color_cells = NULL,
  milestones = NULL,
  milestone_percentages = trajectory$milestone_percentages,
  color_milestones = c("auto", "given", "cubeHelix", "Set3", "rainbow"),
  grouping = NULL,
  groups = NULL,
  scale = dynutils::scale_quantile,
  column_minor_gap = unit(1, "mm"),
  column_major_gap = unit(3, "mm"),
  cell_annotation = NULL,
  cell_info = dataset$cell_info,
  top_annotation = list(),
  plot_milestone_percentages = c("top", "bottom", "none"),
  plot_milestone_network = c("top", "bottom", "none"),
  plot_milestones = c("point", "none", "label"),
  milestone_network_orientation = case_when(first(plot_milestone_network) == "bottom" ~
    "bottom", TRUE ~ "top"),
  milestone_network_arrow = ggplot2::arrow(length = ggplot2::unit(0.3, "cm"), type =
    "closed"),
  plot_velocity = case_when(!is.null(dataset$velocity) ~ "top", TRUE ~ "none"),
  velocity_each = as.integer(length(trajectory$cell_ids)/100),
  heatmap_legend_param = list(direction = "horizontal", border = "#333333", at = c(0,
    1), labels = c("low", "high"), legend_width = unit(3, "cm"))
)
}
\arguments{
\item{features_oi}{The features of interest, either the number of features or a vector giving the names of the different features}

\item{clust}{The method to cluster the features, or a hclust object}

\item{scale}{Whether to rescale the expression, can be a function or boolean}

\item{features_labels}{The labels for the features. If NULL, will use either the names of the features_oi (if exists) or the features_oi themselves. If not NULL, this should be a named character vector providing the label of a features given its feature_id}
}
\description{
Plot the expression across a trajectory in a heatmap
}
\examples{
data(example_linear)
plot_heatmap(example_linear)

data(example_bifurcating)
plot_heatmap(example_bifurcating)

}
\keyword{plot_trajectory}
