% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot the task as a heatmap}
\usage{
plot_heatmap(task,
  clust = hclust(as.dist(correlation_distance(t(task$counts))), method =
  "ward.D2"), margin = 0.02)
}
\arguments{
\item{task}{The task}

\item{clust}{The clustering of the genes as a `clust` object}
}
\description{
Plot the task as a heatmap
}
