% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_connections.R
\name{plot_onedim}
\alias{plot_onedim}
\title{Plot onedim}
\usage{
plot_onedim(task, milestone_network = task$milestone_network,
  color_cells = c("auto", "invisible", "positions", "grouping", "gene"),
  grouping_assignment = NULL, groups = NULL, gene_oi = NULL,
  expression_source = "expression", orientation = 1, margin = 0.05,
  plotdata = make_connection_plotdata(milestone_network, orientation =
  orientation, margin = margin))
}
\arguments{
\item{task}{A trajectory or milestone_network}

\item{milestone_network}{Optional, the milestone network}

\item{color_cells}{How to color the cells}

\item{grouping_assignment}{Tibble containing the assignment of cells to groups of cells}

\item{groups}{Tibble containing information of the cell groups}

\item{gene_oi}{Gene to plot expression}

\item{expression_source}{Source of the gene expression, defaults to `expression`}

\item{orientation}{The orientation of the onedim between milestones}

\item{margin}{Relative distance between milestones}

\item{plotdata}{Optional, The plotting data as generated by make_connection_plotdata}
}
\description{
Plot onedim
}
