% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_onedim.R
\name{plot_onedim}
\alias{plot_onedim}
\title{Plot onedim}
\usage{
plot_onedim(task, color_cells = c("auto", "invisible", "positions",
  "grouping", "gene", "milestone", "pseudotime"),
  milestone_network = task$milestone_network,
  progressions = task$progressions, grouping_assignment = NULL,
  groups = NULL, gene_oi = NULL, pseudotime = NULL,
  expression_source = "expression", color_milestones = c("auto", "given",
  names(milestone_palette_list)), milestones = NULL,
  milestone_percentages = NULL, orientation = 1, margin = 0.05,
  plotdata = make_connection_plotdata(milestone_network, orientation =
  orientation, margin = margin), quasirandom_width = 0.2)
}
\arguments{
\item{task}{A trajectory or milestone_network}

\item{color_cells}{How to color the cells}

\item{milestone_network}{Optional, the milestone network}

\item{grouping_assignment}{Tibble containing the assignment of cells to groups of cells}

\item{groups}{Tibble containing information of the cell groups}

\item{gene_oi}{Gene to plot expression}

\item{pseudotime}{The pseudotime}

\item{expression_source}{Source of the gene expression, defaults to `expression`}

\item{color_milestones}{How to color the cells}

\item{milestones}{Tibble containing the `milestone_id` and a `color` for each milestone}

\item{milestone_percentages}{The milestone percentages}

\item{orientation}{The orientation of the onedim between milestones}

\item{margin}{Relative distance between milestones}

\item{plotdata}{Optional, The plotting data as generated by make_connection_plotdata}

\item{quesirandom_width}{The width of the quasirandom cell spreading}
}
\description{
Plot onedim
}
