% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_onedim.R
\name{plot_onedim}
\alias{plot_onedim}
\title{Plot onedim}
\usage{
plot_onedim(
  trajectory,
  color_cells = c("auto", "none", "grouping", "feature", "milestone", "pseudotime"),
  grouping = NULL,
  groups = NULL,
  feature_oi = NULL,
  pseudotime = NULL,
  expression_source = "expression",
  color_milestones = c("auto", "given", "cubeHelix", "Set3", "rainbow"),
  milestones = NULL,
  milestone_percentages = NULL,
  alpha_cells = 1,
  size_cells = 2.5,
  border_radius_percentage = 0.1,
  orientation = 1,
  margin = 0.05,
  linearised = linearise_cells(trajectory, margin, one_edge = TRUE),
  quasirandom_width = 0.2,
  plot_cells = TRUE,
  label_milestones = dynwrap::is_wrapper_with_milestone_labelling(trajectory)
)
}
\arguments{
\item{trajectory}{The trajectory}

\item{color_cells}{How to color the cells}

\item{grouping}{The grouping of the cells}

\item{groups}{Tibble containing information of the cell groups}

\item{feature_oi}{feature to plot expression}

\item{pseudotime}{The pseudotime}

\item{expression_source}{Source of the feature expression, defaults to `expression`}

\item{color_milestones}{How to color the cells}

\item{milestones}{Tibble containing the `milestone_id` and a `color` for each milestone}

\item{milestone_percentages}{The milestone percentages}

\item{alpha_cells}{The alpha of the cells}

\item{size_cells}{The size of the cells}

\item{border_radius_percentage}{The fraction of the radius that is used for the border}

\item{orientation}{Whether to plot the connections in the top (1) or bottom (-1)}

\item{margin}{The margin to add}

\item{linearised}{The linearised milestone network and progressions}

\item{quasirandom_width}{The width of the quasirandom cell spreading}

\item{plot_cells}{Whether to plot the cells}

\item{label_milestones}{How to label the milestones. Can be TRUE (in which case the labels within the trajectory will be used), "all" (in which case both given labels and milestone_ids will be used), a named character vector, or FALSE}
}
\description{
Plot onedim
}
\examples{
data(example_linear)
plot_onedim(example_linear)
plot_onedim(example_linear, label_milestones = TRUE)

data(example_tree)
plot_onedim(example_tree)
}
\keyword{plot_trajectory}
