% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_strip.R
\name{plot_strip}
\alias{plot_strip}
\title{Plot strip}
\usage{
plot_strip(traj1, traj2, margin = 0.05, reorder = TRUE)
}
\arguments{
\item{traj1}{The first trajectory}

\item{traj2}{The second traj}

\item{margin}{A margin between trajectory segments.}

\item{reorder}{Whether to reorder}
}
\value{
A scatterplot comparison ggplot of two linearised trajectories.
}
\description{
Plot strip
}
\examples{
data(example_bifurcating)
plot_strip(example_bifurcating, example_bifurcating)
}
\keyword{compare_trajectory}
