% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_topology.R
\name{plot_topology}
\alias{plot_topology}
\title{Plot the topology of a trajectory}
\usage{
plot_topology(traj, color_milestones = c("auto", "given", "cubeHelix",
  "Set3", "rainbow"), milestones = NULL, layout = NULL)
}
\arguments{
\item{traj}{The trajectory}

\item{color_milestones}{How to color the cells}

\item{milestones}{Tibble containing the `milestone_id` and a `color` for each milestone}

\item{layout}{The type of layout to create.}
}
\description{
Plot the topology of a trajectory
}
\examples{
data(example_disconnected)
plot_topology(example_disconnected)

data(example_tree)
plot_topology(example_tree)
}
\keyword{plot_trajectory}
