% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_topology.R
\name{plot_topology}
\alias{plot_topology}
\title{Plot the topology of a trajectory}
\usage{
plot_topology(
  trajectory,
  color_milestones = c("auto", "given", "cubeHelix", "Set3", "rainbow"),
  milestones = NULL,
  layout = NULL,
  arrow = grid::arrow(type = "closed", length = unit(0.4, "cm"))
)
}
\arguments{
\item{trajectory}{The trajectory}

\item{color_milestones}{How to color the cells}

\item{milestones}{Tibble containing the `milestone_id` and a `color` for each milestone}

\item{layout}{The type of layout to create. Either a valid string, a
function, a matrix, or a data.frame (see Details)}

\item{arrow}{The type and size of arrow in case of directed trajectories. Set to NULL to remove arrow altogether.}
}
\description{
Plot the topology of a trajectory
}
\examples{
data(example_disconnected)
plot_topology(example_disconnected)

data(example_tree)
plot_topology(example_tree)
}
\keyword{plot_trajectory}
