% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_topology.R
\name{plot_topology}
\alias{plot_topology}
\title{Plotting the topology of a trajectory}
\usage{
plot_topology(task, color_milestones, milestones)
}
\arguments{
\item{task}{The task}

\item{color_milestones}{How to color the cells}

\item{milestones}{Tibble containing the `milestone_id` and a `color` for each milestone}
}
\description{
Plotting the topology of a trajectory
}
